+++ Core/NWNXCore.cpp
@@ -13,6 +13,7 @@
 #include "API/CExoAliasList.hpp"
 #include "API/CVirtualMachine.hpp"
 #include "API/CExoStringList.hpp"
+#include "API/CScriptCompiler.hpp"
 #include "Platform/ASLR.hpp"
 #include "Platform/Debug.hpp"
 #include "Services/Config/Config.hpp"
@@ -233,15 +234,16 @@ void NWNXCore::InitialVersionCheck()
 
         if (version != NWNX_TARGET_NWN_BUILD || revision != NWNX_TARGET_NWN_BUILD_REVISION)
         {
-            std::fprintf(stderr, "NWNX: Expected build version %u revision %u, got build version %u revision %u.",
+            std::fprintf(stderr, "NWNX: Expected build version %u revision %u, got build version %u revision %u.\n",
                                       NWNX_TARGET_NWN_BUILD, NWNX_TARGET_NWN_BUILD_REVISION, version, revision);
+            std::fprintf(stderr, "NWNX: Will terminate. Please use the correct NWNX build for your game version.\n");
             std::fflush(stderr);
-            std::abort();
+            std::exit(1);
         }
     }
     else
     {
-        std::fprintf(stderr, "NWNX: Could not determine build version.");
+        std::fprintf(stderr, "NWNX: Could not determine build version.\n");
         std::fflush(stderr);
         std::abort();
     }
@@ -355,7 +357,7 @@ void NWNXCore::InitialSetupResourceDirectory()
 
 void NWNXCore::InitialSetupCommands()
 {
-    m_services->m_commands->RegisterCommand("runscript", [](std::string& args)
+    m_services->m_commands->RegisterCommand("runscript", [](std::string&, std::string& args)
     {
         if (Globals::AppManager()->m_pServerExoApp->GetServerMode() != 2)
             return;
@@ -367,7 +369,7 @@ void NWNXCore::InitialSetupCommands()
         }
     });
 
-    m_services->m_commands->RegisterCommand("eval", [](std::string& args)
+    m_services->m_commands->RegisterCommand("eval", [](std::string&, std::string& args)
     {
         if (Globals::AppManager()->m_pServerExoApp->GetServerMode() != 2)
             return;
@@ -376,11 +378,14 @@ void NWNXCore::InitialSetupCommands()
         {
             LOG_INFO("Executing console command: 'eval' with args: %s", args);
             bool bWrapIntoMain = args.find("void main()") == std::string::npos;
-            Globals::VirtualMachine()->RunScriptChunk(args, 0, true, bWrapIntoMain);
+            if (Globals::VirtualMachine()->RunScriptChunk(args, 0, true, bWrapIntoMain))
+            {
+                LOG_ERROR("Failed to run console command 'eval' with error: %s", Globals::VirtualMachine()->m_pJitCompiler->m_sCapturedError.CStr());
+            }
         }
     });
 
-    m_services->m_commands->RegisterCommand("evalx", [](std::string& args)
+    m_services->m_commands->RegisterCommand("evalx", [](std::string&, std::string& args)
     {
         if (Globals::AppManager()->m_pServerExoApp->GetServerMode() != 2)
             return;
@@ -403,11 +408,14 @@ void NWNXCore::InitialSetupCommands()
         {
             LOG_INFO("Executing console command: 'evalx' with args: %s", args);
             std::string script = nwnxHeaders + (args.find("void main()") == std::string::npos ? "void main() { " + args + " }" : args);
-            Globals::VirtualMachine()->RunScriptChunk(script, 0, true, false);
+            if (Globals::VirtualMachine()->RunScriptChunk(script, 0, true, false))
+            {
+                LOG_ERROR("Failed to run console command 'evalx' with error: %s", Globals::VirtualMachine()->m_pJitCompiler->m_sCapturedError.CStr());
+            }
         }
     });
 
-    m_services->m_commands->RegisterCommand("loglevel", [](std::string& args)
+    m_services->m_commands->RegisterCommand("loglevel", [](std::string&, std::string& args)
     {
         if (!args.empty())
         {
@@ -440,7 +448,7 @@ void NWNXCore::InitialSetupCommands()
         }
     });
 
-    m_services->m_commands->RegisterCommand("logformat", [](std::string& args)
+    m_services->m_commands->RegisterCommand("logformat", [](std::string&, std::string& args)
     {
         if (args.find("timestamp") != std::string::npos)
             Log::SetPrintTimestamp(args.find("notimestamp") == std::string::npos);
@@ -458,7 +466,6 @@ void NWNXCore::InitialSetupCommands()
                  Log::GetPrintTimestamp(), Log::GetPrintDate(), Log::GetPrintPlugin(),
                  Log::GetPrintSource(), Log::GetColorOutput(), Log::GetForceColor());
     });
-
 }
 
 void NWNXCore::UnloadPlugins()
@@ -513,8 +520,8 @@ void NWNXCore::Shutdown()
 
 void NWNXCore::CreateServerHandler(CAppManager* app)
 {
-    g_core->InitialVersionCheck();
     InitCrashHandlers();
+    g_core->InitialVersionCheck();
 
     g_core->m_services = g_core->ConstructCoreServices();
     g_core->m_coreServices = g_core->ConstructProxyServices(NWNX_CORE_PLUGIN_NAME);
@@ -579,11 +586,12 @@ void NWNXCore::DestroyServerHandler(CAppManager* app)
         }
     }
 
+    auto hook = g_core->m_services->m_hooks->FindHookByAddress(Functions::_ZN11CAppManager13DestroyServerEv);
+    hook->CallOriginal<void>(app);
+
     LOG_NOTICE("Shutting down NWNX.");
     g_core->Shutdown();
 
-    // At this point, the hook has been reset. We should call the original again to let NWN carry on.
-    app->DestroyServer();
     RestoreCrashHandlers();
 }