+++ Plugins/Events/Events/ClientEvents.hpp
@@ -14,6 +14,7 @@ public:
 private:
     static void RemovePCFromWorldHook(bool, CServerExoAppInternal*, CNWSPlayer*);
     static int32_t SendServerToPlayerCharListHook(CNWSMessage*, CNWSPlayer*);
+    static int32_t OnServerCharacterSave(CNWSPlayer*, int32_t);
     static int32_t CheckStickyPlayerNameReservedHook(
             CServerExoApp*,
             CExoString*,