+++ Plugins/Experimental/Experimental.hpp
@@ -5,7 +5,6 @@
 namespace Experimental {
 
 class SuppressPlayerLoginInfo;
-class AsyncLogFlush;
 
 class Experimental : public NWNXLib::Plugin
 {
@@ -15,7 +14,6 @@ public:
 
 private:
     std::unique_ptr<SuppressPlayerLoginInfo> m_SuppressPlayerLoginInfo;
-    std::unique_ptr<AsyncLogFlush> m_AsyncLogFlush;
 };
 
 }