+++ Plugins/Tweaks/Tweaks.hpp
@@ -19,7 +19,6 @@ class StringToIntBaseToAuto;
 class DeadCreatureFiresOnAreaExit;
 class PreserveActionsOnDMPossess;
 class FixGreaterSanctuaryBug;
-class FixScrollLearningBug;
 class ItemChargesCost;
 class AdjustCasterLevel;
 
@@ -44,7 +43,6 @@ private:
     std::unique_ptr<DeadCreatureFiresOnAreaExit> m_DeadCreatureFiresOnAreaExit;
     std::unique_ptr<PreserveActionsOnDMPossess> m_PreserveActionsOnDMPossess;
     std::unique_ptr<FixGreaterSanctuaryBug> m_FixGreaterSanctuaryBug;
-    std::unique_ptr<FixScrollLearningBug> m_FixScrollLearningBug;
     std::unique_ptr<ItemChargesCost> m_ItemChargesCost;
     std::unique_ptr<AdjustCasterLevel> m_AdjustCasterLevel;
 };